test for create_account
