+++ src/operation.rs
@@ -663,45 +663,3 @@ fn convert_xdr_signer_key_to_object(
         _ => panic!("Invalid Type"),
     }
 }
-
-#[cfg(test)]
-mod tests {
-
-    use crate::{
-        account::Account,
-        keypair::{self, Keypair},
-    };
-    use keypair::KeypairBehavior;
-    use xdr::{Int64, OperationBody, ReadXdr};
-
-    use super::*;
-
-    #[test]
-    fn create_account_op_test() {
-        let destination = "GCEZWKCA5VLDNRLN3RPRJMRZOX3Z6G5CHCGSNFHEYVXM3XOJMDS674JZ";
-        let destination_hex =
-            hex!("899b2840ed5636c56ddc5f14b23975f79f1ba2388d2694e4c56ecdddc960e5ef");
-        // println!("Destination hex {:?}", destination_hex);
-        let starting_balance = 1000 * ONE;
-
-        let op = Operation::new()
-            .create_account(destination, starting_balance)
-            .unwrap();
-
-        let op = xdr::Operation::to_xdr(&op, xdr::Limits::none()).unwrap();
-        let op_from = xdr::Operation::from_xdr(op.as_slice(), xdr::Limits::none())
-            .unwrap()
-            .body;
-
-        if let OperationBody::CreateAccount(op) = &op_from {
-            assert_eq!(op.starting_balance, 1000);
-            let mut result: [u8; 32] = Default::default();
-            result[..32].clone_from_slice(&destination_hex);
-            let key = Keypair::new(Some(result), None).unwrap();
-            let val = key.xdr_public_key();
-            assert_eq!(op.destination.0, val);
-        } else {
-            panic!("op is not the type expected");
-        }
-    }
-}
